refactor(value): Moved persistense directory into infrastructure
